requirements.txt
Pillow>=9.0.0
pyautogui>=0.9.54
pywin32>=304
psutil>=5.9.0